android/requirements_android.txt
kivy
pyjnius
# pyjnius  # Décommente pour la synthèse vocale réelle 